gh-54 Link to new ECL Watch pages.
Added links to new result and graph web pages.

Fixes gh-54
