+++ org.hpccsystems.eclide/src/org/hpccsystems/eclide/ui/viewer/platform/GraphFolderView.java
@@ -6,14 +6,14 @@ import java.util.ArrayList;
 import java.util.Observable;
 import java.util.Observer;
 
-import org.hpccsystems.internal.data.Result;
+import org.hpccsystems.internal.data.Graph;
 import org.hpccsystems.internal.data.Workunit;
 import org.hpccsystems.internal.ui.tree.ItemView;
 
-public class ResultFolderView extends FolderItemView implements Observer {
+class GraphFolderView extends FolderItemView implements Observer  {
 	Workunit workunit;
 
-	ResultFolderView(TreeItemOwner treeViewer, PlatformBaseView parent, Workunit wu) {
+	GraphFolderView(TreeItemOwner treeViewer, PlatformBaseView parent, Workunit wu) {
 		super(treeViewer, parent, wu.getPlatform());
 		workunit = wu;
 		workunit.addObserver(this);
@@ -21,19 +21,19 @@ public class ResultFolderView extends FolderItemView implements Observer {
 
 	@Override
 	public String getText() {
-		return "Results";
+		return "Graphs";
 	}
 
 	@Override
 	public URL getWebPageURL() throws MalformedURLException {
-		return platform.getURL("esp/files", "ECLPlaygroundResults.htm", "Wuid=" + workunit.getWuid());
+		return platform.getURL("esp/files", "WUGraph.htm", "Wuid=" + workunit.getWuid());
 	}
 
 	@Override
 	public void refreshChildren() {
 		ArrayList<Object> retVal = new ArrayList<Object>();
-		for(Result r : workunit.getResults()) {
-			retVal.add(new ResultView(treeViewer, this, platform, r));
+		for(Graph g : workunit.getGraphs()) {
+			retVal.add(new GraphView(treeViewer, this, platform, g));
 		}
 		children.set(retVal.toArray(new ItemView[0]));
 	}
@@ -42,7 +42,7 @@ public class ResultFolderView extends FolderItemView implements Observer {
 	public void update(Observable arg0, Object arg1) {
 		if (arg1 instanceof Workunit.Notification) {
 			switch ((Workunit.Notification)arg1){
-			case RESULTS:
+			case GRAPHS:
 				refresh();
 			}
 		}
